pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [
    "setuptools >= 69",
    "wheel >= 0.42",
    "pip >= 23.2.1"
]

[project]
name = "yamldb"
dynamic = ["version"]
description = "A file based database using yaml as file format"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}

authors = [
  {name = "Gregor von Laszewski", email = "laszewski@gmail.com"}
]

maintainers = [
  {name = "Gregor von Laszewski", email = "laszewski@gmail.com"}
]

keywords = ["helper library", "cloudmesh"]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Environment :: Other Environment",
    "Environment :: Plugins",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows :: Windows 10",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Database",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: User Interfaces",
    "Topic :: System",
    "Topic :: System :: Distributed Computing",
    "Topic :: System :: Shells",
    "Topic :: Utilities",
]

dependencies = [
  "cloudmesh-ai-common",
  "jmespath",
  "oyaml"
]

[project.urls]
Homepage = "https://github.com/cloudmesh/yamldb"
Documentation = "https://github.com/cloudmesh/yamldb/blob/main/README.md"
Repository = "https://github.com/cloudmesh/yamldb.git"
Issues = "https://github.com/cloudmesh/yamldb/issues"
Changelog = "https://github.com/cloudmesh/yamldb/blob/main/CHANGELOG.md"


[tool.setuptools.dynamic]
version = {file = "VERSION"}

[tool.setuptools.packages.find]
where = ["src/"]
# include = ["yamldb"]
